backend/requirements.txt
psycopg2-binary==2.8.3
Django==2.1.7
django-cors-headers==2.4.0
djangorestframework==3.9.1
Pillow==6.1.0
django-extensions==2.2.3
